fix(elreal): address PR #1082 review (scale width, to_triple, denorm_min, Codacy)
CodeRabbit (3 Major) + Codacy:
- elreal::scale() narrowed the unbounded integer<256> exponent to int -> could
  overflow. Widen to int64_t (vast headroom for the host-FP attribute domain it
  feeds; full width stays reachable via block::exponent()).
- manipulators to_triple() emitted the full VALUE as the third field; the
  contract is (sign, scale, significand). Emit significand = ldexp(approx,-scale)
  in [1,2). Strengthened the test to assert to_triple(1/8) == "(+, -3, 1)".
- numeric_limits denorm_min(): has_denorm = denorm_absent, so the C++20 contract
  requires the minimum positive NORMALISED value -> return min(). Added a test.
- Codacy knownConditionTrueFalse (lim::radix != 2 etc. fold): the compile-time
  numeric_limits fields are now checked with static_assert (idiomatic for
  constexpr members; no folded runtime comparisons); runtime checks remain only
  for the value factories.

el_api_attributes + el_api_api + el_api_lazy PASS on gcc + clang; cppcheck no
longer reports knownConditionTrueFalse; ASCII clean.

Relates to #1079

+++ include/sw/universal/number/elreal/elreal_impl.hpp
@@ -21,6 +21,7 @@
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <cstddef>
+#include <cstdint>
 #include <string>
 #include <vector>
 #include <type_traits>
@@ -183,9 +184,13 @@ public:
         auto bl = _value.take(1);
         return (!bl.empty() && bl.front().sign() < 0) ? -1 : 1;
     }
-    int scale() const noexcept {
+    // int64_t (not int): elreal carries an unbounded integer<256> exponent, so a
+    // narrow int cast could overflow. The host-FP attribute surface this feeds
+    // (ldexp / significand) is bounded by FpType's exponent range, for which int64_t
+    // has vast headroom; the full-width exponent stays reachable via block::exponent().
+    int64_t scale() const noexcept {
         auto bl = _value.take(1);
-        return bl.empty() ? 0 : static_cast<int>(bl.front().exponent());
+        return bl.empty() ? 0 : static_cast<int64_t>(bl.front().exponent());
     }
 
 private:
